rename tissue_sample validator and add new validator for sample duplication and consistency
